script/requirements.txt
fastapi==0.114.2
uvicorn[standard]==0.30.6
pillow==10.4.0
imagehash==4.3.1
pydantic==2.9.2 